requirements.txt
streamlit
psycopg2-binary
opencv-python
numpy
Pillow
pyzbar
